Add files via upload
This is terraform workflow. we need to understand how the command work behind the scene
WE are terraform admin, we have our laptop or a local desktop, we install our Terraform CLI we have our CLI . W e get the CLI by typing Terraform --version and from there we have our CLI
We  add some files  like main.tf for example, we put some code in our main.tf file with the provider block and terraform setting block
We type the first command  Terraform init or packer init or git init
-In this case we will type terraform init , so terraform will look inside that file if we have specified a terraform block, because into our terraform block, we will specify a required provider, and terraform will go into the required registry and download the required provider and store it into our local machine and do no further action. `` All terraform provider are public,, all provider write their own provider code that's how  we use to query the AWS or the cloud provider API to deploy in AWS, so when we want to deploy a in AWS Terraform will use the provider to know how to deploy`` . All provider have version, so we have to specify our version to avoid terraform to use the latest version that might come with new features 
-Terraform init main goal is to download the latest provider
-Terraform validate  is used to check if all our configuration files are well organized and look good
-Terraform fmt , is a format command that check if all our terrafrom tf. files code follow the hcl language and it will format it
-Terraform plan, When we type it the first plan, terraform will create a 2 files called terraform.tfstate and terraform.tfstate backup. Terraform.tf state will recall everything terrafom will create in AWS , and the terrafrom state backup is just the backup of our terraform. tf state and it is store into our local machine. ( the second time we type terraform plan, terraform will not create a terraform.tf state it will just check if the existing one looks goods)
The second time  we type terraform plan, terraform will ckeck if we have a folder that hold the key call ~/.AWS  it is the same like ~/.Kube/config that we use to access the cluster. When terraform find the .AWS key it will check if we have the right access and the right secret access key to access the aws API
When Terraform find those access keys it will try to authenticate through AWS, if those keys are not correct, we will have permission denied. If the keys are correct terraform will authenticate through  AWS using IAM (identity access management). 
When it authenticate terraform will check the policy attached to our account, such as are we able to create an AWS instance?  If our policy do not give not enough permissions, we will have permissions denied , even though we have access to AWS , So when terraform check all our policy and realized that we have the proper permission , that's is how it will run terraform plan.
-To run terraform plan , terraform will check the desired state into our statefile and and the current state in AWS. It's compare to argocd in kubernetes. So if we run terraform plan and it happenend that there is an ec2 instance non existent in the current state in aws , but exist in the desire state, terraform will create it. Terraform does not have the sync in feature like argocd , which is able to detect the changes in both state and fix them and if auto sync is enable, it will deploy it. In terraform, we actually have to type terraform plan and terrafom apply to fix the issue, Terraform does not detect the changes like in argo cd
-terraform plan compare the desire state to our current state, if there any resource added into our configuration file, terraform plan  will let us know  and ask if we want to make changes and deploy it so that both states (desire and current  matches0, when we type, that is how terraform apply will make the changes and both states file will be the same
- so the second time that we type terraform plan, if there is no difference between the desire state and the current state, terraform will send us a message saying that we are UpToDate, and if there is any changes, terraform will ask us if we want to sync it and terraform apply will make the sync in